\name{SynClustR-package}
\alias{SynClustR-package}
\alias{SynClustR}
\docType{package}
\title{\packageTitle{SynClustR}}
\description{\packageDescription{SynClustR}}
\author{\packageAuthor{SynClustR}}
\section{Maintainer}{\packageMaintainer{SynClustR}}
\references{GSL: GNU Scientific Library: \url{http://www.gnu.org/software/gsl/}}
\keyword{ package }

